src/main/java/com/example/UnitTestExempel1/OrderService.java
package com.example.UnitTestExempel1;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class OrderService {

    private final OrderRepository orderRepository;

    @Autowired
    public OrderService(OrderRepository orderRepository) {
        this.orderRepository = orderRepository;
    }

    public Order createOrder(Order order){
        return orderRepository.save(order);
    }

    public Order findOrderById(Long id){
        return orderRepository.findById(id).orElse(null);
    }
}
